Rename library to `ShopifyCheckoutKit`.
