+++ Package.swift
@@ -4,15 +4,15 @@
 import PackageDescription
 
 let package = Package(
-    name: "ShopifyCheckout",
+    name: "ShopifyCheckoutKit",
     platforms: [
         .iOS(.v13)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "ShopifyCheckout",
-            targets: ["ShopifyCheckout"])
+            name: "ShopifyCheckoutKit",
+            targets: ["ShopifyCheckoutKit"])
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
@@ -23,11 +23,11 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "ShopifyCheckout",
+            name: "ShopifyCheckoutKit",
             dependencies: []),
         .testTarget(
-            name: "ShopifyCheckoutTests",
-            dependencies: ["ShopifyCheckout"],
+            name: "ShopifyCheckoutKitTests",
+            dependencies: ["ShopifyCheckoutKit"],
             plugins: [
                 .plugin(name: "SwiftLint", package: "SwiftLintPlugin")
             ])
