+++ Samples/MobileBuyIntegration/MobileBuyIntegration/AppConfiguration.swift
@@ -21,7 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-import ShopifyCheckout
+import ShopifyCheckoutKit
 
 public struct AppConfiguration {
 	public var useVaultedState: Bool = false
