+++ Samples/MobileBuyIntegration/MobileBuyIntegration/ProductViewController.swift
@@ -23,7 +23,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SO
 
 import Buy
 import UIKit
-import ShopifyCheckout
+import ShopifyCheckoutKit
 
 class ProductViewController: UIViewController {
