+++ Samples/README.md
@@ -1,6 +1,6 @@
 # Sample Projects
 
-In this directory you will find sample projects which implement the `ShopifyCheckout` library in various ways.
+In this directory you will find sample projects which implement the `ShopifyCheckoutKit` library in various ways.
 
 Each project directory contains a `Storefront.xcconfig.example` file. Simply rename it to `Storefront.xcconfig` and update the contained values to match your Shopify storefront.
 
@@ -8,8 +8,8 @@ Each project directory contains a `Storefront.xcconfig.example` file. Simply ren
 
 ### MobileBuyIntegration
 
-This project demonstrates how to use the [Mobile Buy SDK](https://github.com/Shopify/mobile-buy-sdk-ios) in conjunction with the `ShopifyCheckout` library.
+This project demonstrates how to use the [Mobile Buy SDK](https://github.com/Shopify/mobile-buy-sdk-ios) in conjunction with the `ShopifyCheckoutKit` library.
 
 ### SimpleAppIntegration
 
-This project demonstrates how to use the `ShopifyCheckout` library with a custom built Storefront GraphQL API client.
+This project demonstrates how to use the `ShopifyCheckoutKit` library with a custom built Storefront GraphQL API client.
