+++ Sources/ShopifyCheckoutKit/Assets.xcassets/Contents.json
