+++ Sources/ShopifyCheckoutKit/Assets.xcassets/spinner.imageset/Contents.json
