+++ Sources/ShopifyCheckoutKit/CheckoutBridge.swift
@@ -29,8 +29,8 @@ enum CheckoutBridge {
 	static let messageHandler = "mobileCheckoutSdk"
 
 	static var applicationName: String {
-		let theme = ShopifyCheckout.configuration.colorScheme.rawValue
-		return "ShopifyCheckoutSDK/\(ShopifyCheckout.version) (\(schemaVersion);\(theme);standard)"
+		let theme = ShopifyCheckoutKit.configuration.colorScheme.rawValue
+		return "ShopifyCheckoutKitSDK/\(ShopifyCheckoutKit.version) (\(schemaVersion);\(theme);standard)"
 	}
 
 	static func decode(_ message: WKScriptMessage) throws -> WebEvent {
