+++ Sources/ShopifyCheckoutKit/CheckoutViewController.swift
@@ -77,7 +77,7 @@ class CheckoutViewController: UIViewController, UIAdaptivePresentationController
 	override public func viewDidLoad() {
 		super.viewDidLoad()
 
-		view.backgroundColor = ShopifyCheckout.configuration.backgroundColor
+		view.backgroundColor = ShopifyCheckoutKit.configuration.backgroundColor
 
 		view.addSubview(checkoutView)
 		NSLayoutConstraint.activate([
